Lift up deriveAddress function
